changed some seed stuff

+++ db/seeds.rb
@@ -5,18 +5,28 @@
 #
 #   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
 #   Character.create(name: 'Luke', movie: movies.first)
-require 'json'
 require 'open-uri'
-Ingredient.destroy_all
+Ingredient.destroy_all if Rails.env.development?
+Cocktail.destroy_all if Rails.env.development?
+puts "Deleted all ingredients and cocktails"
 
-open("https://www.thecocktaildb.com/api/json/v1/1/list.php?i=list") do |ingredients|
-  data = []
-  ingredients.read.each_line do |ingredient|
-    @item = JSON.parse(ingredient)
-    object = {
-      "name": @item["drinks.strIngredient1"]
-    }
-    data << object
-  end
-  Ingredient.create!(data)
+ingredients = JSON.parse(open("https://www.thecocktaildb.com/api/json/v1/1/list.php?i=list").read)
+# puts ingredients
+ingredients["drinks"].each do |ingredient|
+  Ingredient.create(name: ingredient["strIngredient1"])
+  puts "Created all cocktails"
 end
+
+
+# open("https://www.thecocktaildb.com/api/json/v1/1/list.php?i=list") do |ingredients|
+#   data = []
+#   ingredients.read.each_line do |ingredient|
+#     @item = JSON.parse("https://www.thecocktaildb.com/api/json/v1/1/list.php?i=list")
+#     object = {
+#       "name": @item["drinks.strIngredient1"]
+#     }
+#     data << object
+#   end
+#   Ingredient.create!(data)
+# end
+#
